refactor: move PokemonDetails
